Allow exposing public_api via get_api_by_name()
This commit changes the required method signature of API class ctors.
The main reason is that the mapping of names to API objects occurs at
the connection level, which is an intermediate level between the level
of the app and the level of the individual API.  A shared_ptr reference
to api_connection_context (a container for state at this intermediate
level) is now passed to API ctor's along with the & reference to the
application.