Added extra parameters to get_discussions_by_* API
When querying a list of posts you can now specify "select_author" and
"select_tags" to white-list the set of authors/tags that you would like
returned. If not specified then it behaves as before.

Additionally you can specify "truncate_body" with a number of characters
and the body contents will be truncated.  The full length of the body
will also be provided in "body_length" field on the comment. This can
inform the user interface whether or not the body has been truncated and
may be in need of fetching.

If truncate_body is omited or set to 0 then the full body will be
returned.

get_state currently does not truncate and/or filter.