Implement multiple beneficiaries. #432
As a comment option, authors should be able to specify a number of
beneficiaries to receive a percentage of the rewards and the percent of
the potential liquid rewards to receive as Steem Power. Unallocated
rewards by default go to the author as they currently do.

Inspired by steemit#773 and supported by the community.

Commit contains fix miscalculation when paying beneficiaries steemit#909 .

As described in steemit#2162, 8 beneficiniries is small,
that is why limit is 64.